requirements (1).txt
cryptography>=39.0.0
PyPDF2>=3.0.0
pypdf>=5.4.0
python-magic>=0.4.27
chardet>=5.0.0
filetype>=1.2.0
PyQt6>=6.8.0
pyjwt>=2.6.0
jwcrypto>=1.4.2
flask-swagger-ui>=4.11.1
flasgger>=0.9.5
flask-cors>=5.0.1
